// frontend/src/components/design-system/input/ButtonComponent.types.ts
export enum Color {
  Neutral = 'neutral',
  Primary = 'primary',
  Secondary = 'secondary',
  White = 'white',
  Facebook = 'facebook',
}

export enum Kind {
  Link = 'link',
  Outlined = 'outlined',
  Filled = 'filled',
  Answer = 'answer',
}

export enum Size {
  Medium = 'medium',
  Small = 'small',
}

export enum Tag {
  Anchor = 'a',
  Button = 'button',
}